feat: add server key setter
