--- fhe_http_python/src/apis/mod.rs
@@ -1,6 +1,0 @@
-pub mod base64;
-pub mod fhe;
-pub mod fhe_ops;
-pub mod http;
-pub mod key_generator;
-pub mod serializer;
