+++ fhe_http_python/src/apis/server_key_setter.rs
@@ -1,0 +1,26 @@
+use fhe_http_core::configs::typing::SerialServerKey;
+use fhe_http_core::fhe_traits::serializable::KeySerializable;
+use fhe_http_core::tfhe::{set_server_key, CompressedServerKey, ServerKey};
+use pyo3::prelude::*;
+
+#[pyclass]
+pub struct ServerKeySetter {}
+
+#[pymethods]
+impl ServerKeySetter {
+    #[new]
+    pub fn new() -> Self {
+        ServerKeySetter {}
+    }
+    fn decompress_server_key(&self, server_key: SerialServerKey) -> Vec<u8> {
+        let compressed_sks: CompressedServerKey =
+            KeySerializable::try_deserialize(&server_key).unwrap();
+        let decompressed_sks = compressed_sks.decompress();
+        decompressed_sks.try_serialize().unwrap().clone()
+    }
+
+    pub fn set_server_key(&self, server_key: SerialServerKey) {
+        let server_key: ServerKey = KeySerializable::try_deserialize(&server_key).unwrap();
+        set_server_key(server_key);
+    }
+}
